Demonstrate sticky header table more clearly in docs
Previously, the example in the documentation for sticky header tables did not
actually show the sticky header behaviour, since the table was not scrollable.
This update increases the number of rows in the table and forces the table's
wrapper to have a specific height, causing the table to become scrollable and
thus demonstrating the sticky header behaviour.