Add missing TOC item on the preview's form elements page
Previously, the section heading for 'Input with appended `<kbd>` was not
appearing on that page's table of contents. This was due to how the headings on
pages were parsed when constructing the TOC.

This change updates the heading. It seems reasonable to simplify the heading,
rather than modify the code that constructs the TOC.