Rename several Theme builder labels
We have updated the heading on the top nav button, the heading inside the theme
settings side panel, and the side panel's save button.

There were a couple of reasons why this change was recommended:
1)  the name 'Settings' in the top nav was not really clear about what _kind_ of
settings it represented
2) There is already a link in the 'Extra' nav menu called 'Settings', and having
both items named the same was potentially confusing.
3) Theme Settings seemed less jargon-y and more straightforward for
non-technical folks to understand than Theme Builder.
4) 'Save' seemed shorter and still simple enough to understand over 'Save
settings'.
5) The settings icon on the save button has been removed as it seemed
superfluous and/or unrelated.