Allow Offcanvas docs page to scroll properly
Previously, The Offcanvas documentation examples were preventing the rest of the
page from scrolling. This was easy enough to fix with a data attribute as per
the Bootstrap documentation. However, it was also necessary to change the
examples so that instead of forcing the Offcanvas elements into their 'visible'
state with a hardcoded css class, they are triggered on page load with an inline
javascript function.